Refactor: consolidate flash-mtp decode metadata and host fixtures
- Rename decode_metadata_device.py -> decode_metadata.py so the on-device
  kernel owns the plain module name, matching every other kernel in the
  directory
- Merge the host-side torch fixture helpers (the former decode_metadata.py
  and rope_tables.py) into a single utils.py; both were fixture-only and
  imported solely from test sections
- Rename build_deepseek_v4_rope_tables -> build_rope_tables
- Lift cache_row_from_table into utils, dropping six identical copies
  across the prefill SWA/HCA/CSA fixtures
- build_decode_metadata now calls build_swa_metadata instead of repeating
  its two spmd loops verbatim, which also leaves a single
  decode_build_swa_metadata name_hint instead of two colliding ones
- Drop build_swa_metadata's unused Out-param return, and collapse
  ori_slot_mapping onto paged_slot_mapping, whose body it duplicated
