+++ models/deepseek_v4_flash_mtp/decode_attention_csa.py
@@ -522,7 +522,7 @@ def golden_attention_csa(tensors):
 
 def build_tensor_specs(start_pos=None):
     import torch
-    from decode_metadata import (
+    from utils import (
         block_table,
         compressed_slot_mapping,
         csa_decode_start_set,
@@ -535,9 +535,9 @@ def build_tensor_specs(start_pos=None):
     )
     from golden import TensorSpec
     from hc_pre import golden_hc_pre
-    from rope_tables import build_deepseek_v4_rope_tables
+    from utils import build_rope_tables
 
-    shared_freqs_cos, shared_freqs_sin = build_deepseek_v4_rope_tables(M, COMPRESS_RATIO, dtype=torch.bfloat16)
+    shared_freqs_cos, shared_freqs_sin = build_rope_tables(M, COMPRESS_RATIO, dtype=torch.bfloat16)
     def round_half_away_from_zero(x):
         return torch.sign(x) * torch.floor(torch.abs(x) + 0.5)
