+++ models/deepseek_v4_flash_mtp/decode_attention_swa.py
@@ -317,7 +317,7 @@ def golden_attention_swa(tensors):
 
 def build_tensor_specs(start_pos=None):
     import torch  # type: ignore[import]
-    from decode_metadata import (
+    from utils import (
         block_table,
         paged_slot_mapping,
         position_ids_from_starts,
@@ -326,9 +326,9 @@ def build_tensor_specs(start_pos=None):
         swa_decode_start_set,
     )
     from golden import TensorSpec
-    from rope_tables import build_deepseek_v4_rope_tables
+    from utils import build_rope_tables
 
-    shared_freqs_cos, shared_freqs_sin = build_deepseek_v4_rope_tables(M, 0, dtype=torch.bfloat16)
+    shared_freqs_cos, shared_freqs_sin = build_rope_tables(M, 0, dtype=torch.bfloat16)
 
     def quant_w_per_output_channel(w):
         amax = w.float().abs().amax(dim=0).clamp_min(INT8_AMAX_EPS)
