+++ models/deepseek_v4_flash_mtp/decode_sparse_attn_hca.py
@@ -767,7 +767,7 @@ def build_tensor_specs(
 ):
     """Build deterministic demo tensors for the merged standalone harness."""
     import torch
-    from decode_metadata import block_table
+    from utils import block_table
     from golden import TensorSpec
 
     cmp_valid = min(get_standalone_cmp_valid(compress_ratio), TOPK - WIN)
