--- models/deepseek_v4_flash_mtp/rope_tables.py
@@ -1,157 +1,0 @@
-# Copyright (c) PyPTO Contributors.
-# This program is free software, you can redistribute it and/or modify it under the terms and conditions of
-# CANN Open Software License Agreement Version 2.0 (the "License").
-# Please refer to the License for details. You may not use this file except in compliance with the License.
-# THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
-# INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
-# See LICENSE in the root of the software repository for the full text of the License.
-# -----------------------------------------------------------------------------------------------------------
-"""Host-side DeepSeek-V4 RoPE/YaRN table generation.
-
-The V4 kernels consume real-valued cos/sin tables instead of the complex
-``freqs_cis`` tensor used by ``model.py``.  This module keeps the same YaRN
-frequency math and exposes tables in the current kernel ABI.
-"""
-
-from __future__ import annotations
-
-import math
-from typing import Any
-
-import torch
-
-
-def _torch_dtype(dtype: torch.dtype | str) -> torch.dtype:
-    if isinstance(dtype, torch.dtype):
-        return dtype
-    normalized = dtype.lower()
-    if normalized in {"bf16", "bfloat16", "torch.bfloat16"}:
-        return torch.bfloat16
-    if normalized in {"fp32", "float32", "torch.float32"}:
-        return torch.float32
-    if normalized in {"fp16", "float16", "torch.float16"}:
-        return torch.float16
-    raise ValueError(f"Unsupported RoPE table dtype: {dtype!r}")
-
-
-def rope_profile_for_compress_ratio(config: Any, compress_ratio: int) -> tuple[float, int]:
-    """Return ``(base_theta, original_seq_len)`` for the two DeepSeek-V4 RoPE profiles."""
-    if compress_ratio:
-        return float(config.compress_rope_theta), int(config.original_max_position_embeddings)
-    return float(config.rope_theta), 0
-
-
-def _linear_ramp_factor(low: int, high: int, dim: int, *, device: torch.device | None = None) -> torch.Tensor:
-    if low == high:
-        high = high + 0.001
-    ramp = (torch.arange(dim, dtype=torch.float32, device=device) - low) / (high - low)
-    return torch.clamp(ramp, 0, 1)
-
-
-def _find_correction_dim(num_rotations: int, dim: int, base: float, max_seq_len: int) -> float:
-    return dim * math.log(max_seq_len / (num_rotations * 2 * math.pi)) / (2 * math.log(base))
-
-
-def _find_correction_range(
-    low_rot: int,
-    high_rot: int,
-    dim: int,
-    base: float,
-    max_seq_len: int,
-) -> tuple[int, int]:
-    low = math.floor(_find_correction_dim(low_rot, dim, base, max_seq_len))
-    high = math.ceil(_find_correction_dim(high_rot, dim, base, max_seq_len))
-    return max(low, 0), min(high, dim - 1)
-
-
-def precompute_freqs_cos_sin(
-    dim: int,
-    seqlen: int,
-    original_seq_len: int,
-    base: float,
-    factor: float,
-    beta_fast: int,
-    beta_slow: int,
-    *,
-    dtype: torch.dtype | str = torch.bfloat16,
-    device: torch.device | str | None = None,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Return real RoPE tables equivalent to ``model.py::precompute_freqs_cis``.
-
-    The returned tensors are shaped ``[seqlen, dim]``.  The first half contains
-    the mathematical ``cos(angle)`` / ``sin(angle)`` values; the second half is a
-    duplicate so kernels can either read ``:dim//2`` directly or use ``j >> 1``
-    frequency duplication over a full-width table.
-    """
-    if dim <= 0 or dim % 2 != 0:
-        raise ValueError(f"RoPE dim must be a positive even integer, got {dim}")
-    if seqlen <= 0:
-        raise ValueError(f"RoPE sequence length must be positive, got {seqlen}")
-
-    out_dtype = _torch_dtype(dtype)
-    out_device = torch.device(device) if device is not None else None
-    half_dim = dim // 2
-
-    inv_freq = 1.0 / (
-        float(base) ** (torch.arange(0, dim, 2, dtype=torch.float32, device=out_device) / dim)
-    )
-    if original_seq_len > 0:
-        low, high = _find_correction_range(beta_fast, beta_slow, dim, float(base), int(original_seq_len))
-        smooth = 1 - _linear_ramp_factor(low, high, half_dim, device=out_device)
-        inv_freq = inv_freq / float(factor) * (1 - smooth) + inv_freq * smooth
-
-    positions = torch.arange(seqlen, dtype=torch.float32, device=out_device)
-    angles = torch.outer(positions, inv_freq)
-    cos_half = torch.cos(angles)
-    sin_half = torch.sin(angles)
-    freqs_cos = torch.cat([cos_half, cos_half], dim=-1).to(out_dtype)
-    freqs_sin = torch.cat([sin_half, sin_half], dim=-1).to(out_dtype)
-    return freqs_cos, freqs_sin
-
-
-def build_deepseek_v4_rope_tables(
-    config: Any,
-    compress_ratio: int,
-    *,
-    max_seq_len: int | None = None,
-    rope_dim: int | None = None,
-    dtype: torch.dtype | str = torch.bfloat16,
-    device: torch.device | str | None = None,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Return ``(freqs_cos, freqs_sin)`` shaped ``[max_seq_len, rope_dim]``."""
-    base, original_seq_len = rope_profile_for_compress_ratio(config, compress_ratio)
-    seq_len = int(max_seq_len if max_seq_len is not None else config.max_position_embeddings)
-    dim = int(rope_dim if rope_dim is not None else config.qk_rope_head_dim)
-
-    return precompute_freqs_cos_sin(
-        dim,
-        seq_len,
-        original_seq_len,
-        base,
-        float(config.rope_factor),
-        int(config.beta_fast),
-        int(config.beta_slow),
-        dtype=dtype,
-        device=device,
-    )
-
-
-def materialize_token_rope_tables(
-    freqs_cos: torch.Tensor,
-    freqs_sin: torch.Tensor,
-    position_ids: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Gather token-local RoPE tables using absolute ``position_ids``."""
-    positions = position_ids.to(device=freqs_cos.device, dtype=torch.long).reshape(-1)
-    return freqs_cos.index_select(0, positions).contiguous(), freqs_sin.index_select(0, positions).contiguous()
-
-
-def materialize_half_rope_tables(
-    freqs_cos: torch.Tensor,
-    freqs_sin: torch.Tensor,
-    positions: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Gather half-width FP32 cos/sin rows for decode submodule fixtures."""
-    cos, sin = materialize_token_rope_tables(freqs_cos, freqs_sin, positions)
-    half_dim = freqs_cos.shape[-1] // 2
-    return cos[:, :half_dim].float().contiguous(), sin[:, :half_dim].float().contiguous()
